requirements.txt
python-binance==1.0.36
python-dotenv==1.2.2
pandas==3.0.1
ta==0.11.0
requests==2.33.0
schedule==1.2.2
